Day 0 - Getting off the ground
The aim was to create a project and a road forward, such was accomplished and the task feels doable. Folder skeleton has been made and README has been updated (will need further updating as I learn more).

Next Step: Learn how to use numpy, pandas and matplotlib to perform EDA and clean data as recommended by discussions on the Kaggle website. After that, perhaps attempt to create a digestible visual by the end of day 2 before beginning work on the NN.