Day 0.75 - Kaggle-to-VSCode assimilation
After doing a lot of research on basic Exploratory Data Analysis, I decided it would be most comfortable for me to create and use a Jupyter Notebook on VSCode.

In setting up this environment, I created a cell that downloads the necessary raw data when run. I could argue it keeps the code more self-contained but it was moreso to satisfy my curiosity as to if it could be done. Though I hope in the future it will abstract a lot of finnicky bits from people who wish to download and run my code.

Details on setup of Kaggle into using my code to download the dataset have been pushed into README.MD.

Actually doing the Exploratory Data Analysis is my next step, hence this is day 0.75.